+++ tests/testthat/test-update.R
@@ -543,3 +543,14 @@ test_that("NAs propagate in update", {
   expect_equal(time_update(ymd("2020-03-03"), year = c(2021, NA), second = c(NA, 10)),
                c(NA_POSIXct_, NA_POSIXct_))
 })
+
+
+test_that("tzone attributes of Dates is preserved", {
+  d <- ymd("2020-01-01")
+  tzone <- "America/New_York"
+  attr(d, "tzone") <- tzone
+  expect_is(time_update(d, month = 2), "Date")
+  expect_is(time_update(d, hour = 2), "POSIXct")
+  expect_identical(time_update(d, month = 2), structure(ymd("2020-02-01"), tzone = tzone))
+  expect_identical(time_update(d, hour = 1), ymd_hms("2020-01-01 01:00:00", tz = tzone))
+})